+++ test/Util/LimitedArrayCacheTest.php
@@ -7,26 +7,26 @@ use PHPUnit\Framework\TestCase;
 
 class LimitedArrayCacheTest extends TestCase
 {
-    public function testSaveAndFetch()
+    public function testSetAndGet()
     {
         $cache = new LimitedArrayCache();
-        $cache->save('test_key', 'value');
-        $this->assertEquals('value', $cache->fetch('test_key'));
+        $cache->set('test_key', 'value');
+        $this->assertEquals('value', $cache->get('test_key'));
     }
 
     public function testDeleteAndContains()
     {
         $cache = new LimitedArrayCache();
-        $cache->save('test_key', 'value');
+        $cache->set('test_key', 'value');
         $cache->delete('test_key');
-        $this->assertFalse($cache->contains('test_key'));
+        $this->assertFalse($cache->has('test_key'));
     }
 
-    public function testSaveMultipleAndFetchMultiple()
+    public function testSetMultipleAndFetchMultiple()
     {
         $cache = new LimitedArrayCache();
-        $cache->saveMultiple(['key1'=>'value1', 'key2'=>'value2']);
-        $values = $cache->fetchMultiple(['key2', 'key1']);
+        $cache->setMultiple(['key1'=>'value1', 'key2'=>'value2']);
+        $values = $cache->getMultiple(['key2', 'key1']);
         $this->assertEquals(['key2'=>'value2', 'key1'=>'value1'], $values);
     }
 
@@ -37,29 +37,29 @@ class LimitedArrayCacheTest extends TestCase
         for($i = 1; $i < 12; $i++) {
             $data['key'.$i] = 'value'.$i;
         }
-        $cache->saveMultiple($data);
-        $this->assertEquals(11, $cache->getStats()[$cache::STATS_KEYS]);
-        $cache->save('key12', 'value12');
-        $this->assertEquals(6, $cache->getStats()[$cache::STATS_KEYS]);
-        $this->assertFalse($cache->contains('key6'));
-        $this->assertTrue($cache->contains('key7'));
+        $cache->setMultiple($data);
+        $this->assertEquals(11, $cache->count());
+        $cache->set('key12', 'value12');
+        $this->assertEquals(6, $cache->count());
+        $this->assertFalse($cache->has('key6'));
+        $this->assertTrue($cache->has('key7'));
     }
 
     public function testExpiration()
     {
         $cache = new LimitedArrayCache();
-        $cache->save('test_key', 'value', 0.5);
-        $this->assertTrue($cache->contains('test_key'));
-        sleep(1);
-        $this->assertFalse($cache->contains('test_key'));
+        $cache->set('test_key', 'value', 1);
+        $this->assertTrue($cache->has('test_key'));
+        sleep(2);
+        $this->assertFalse($cache->has('test_key'));
     }
 
-    public function testFlushAll()
+    public function testClear()
     {
         $cache = new LimitedArrayCache();
-        $cache->save('test_key', 'value');
-        $cache->flushAll();
-        $this->assertFalse($cache->contains('test_key'));
-        $this->assertEquals(0, $cache->getStats()[$cache::STATS_KEYS]);
+        $cache->set('test_key', 'value');
+        $cache->clear();
+        $this->assertFalse($cache->has('test_key'));
+        $this->assertEquals(0, $cache->count());
     }
 }