+++ CHANGELOG.md
@@ -5,9 +5,10 @@ Breaking Changes
 ----------------
 
 * Now Requires a PSR-11 container for data object construction
-* Requires PHP 8 or above
+* Requires PHP 8.0 or above
 * Depends on Doctrine/DBAL 3.x
 * Phpdoc annotations have been replaced with DbTable and IdColumn attributes
+* Cache uses PSR-16 Simple Cache interface rather than requiring Doctrine's Cache implementation
 
 Updates
 -------